12 Jan 2024, 01:12

+++ squareRootUsing binary search.cpp
@@ -3,7 +3,7 @@
 
 using namespace std;
 
-int squareRoot(int num,int start,int end,int mid)
+int squareRoot(int num,int start,int end,long long int mid)
 {
     mid=mid/2;
     if(mid*mid==num){
